+++ e2e-tests/jest.config.ts
@@ -1,22 +1,23 @@
-import { config } from 'dotenv';
+import { config as dotenvConfig } from 'dotenv';
+import type { Config as JestConfig } from 'jest';
 
-config({
+dotenvConfig({
   path: '.env',
 });
 
-export default {
-  collectCoverage: true,
-  testTimeout: 360000,
+const config: JestConfig = {
+  collectCoverage: false,
+  testTimeout: process.env.RECORD_FIXTURES === 'true' ? 240 * 1000 : 10 * 1000,
   testEnvironment: 'node',
-  testMatch: ['**/*.test.ts'],
+  testMatch: ['**/*.test.ts', '!**/test-applications/**/*.test.ts'],
   transform: {
     '^.+\\.tsx?$': 'ts-jest',
   },
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
-  globals: {
-    'ts-jest': {
-      tsconfig: 'tsconfig.json',
-    },
-  },
   verbose: true,
+  setupFilesAfterEnv: ['<rootDir>/mocks/setup.ts'],
+  globalSetup: '<rootDir>/global-setup.ts',
+  globalTeardown: '<rootDir>/global-teardown.ts',
 };
+
+export default config;